+++ packages/ui/src/components/w3m-wallet-button/index.ts
@@ -39,7 +39,7 @@ export class W3mWalletButton extends LitElement {
         <div>
           <w3m-wallet-image walletId=${this.walletId} .src=${this.src}></w3m-wallet-image>
           <w3m-text variant="xsmall-normal">
-            ${this.label ?? UiUtil.getWalletFirstName(this.name)}
+            ${this.label ?? UiUtil.getWalletName(this.name, true)}
           </w3m-text>
 
           ${this.sublabelTemplate()}
